Choosing the Right Ultrasound Imaging Center for Your Pregnancy at Butler Medical Imaging. (08) 9544 3999
Pregnancy is a time of excitement and anticipation, as you prepare to bring a new life into the world. But it can also bring its fair share of questions and concerns. One of the most common is whether ultrasound sonography is necessary and if these high-frequency waves are safe for the baby.

What is an Ultrasound Scan?

An ultrasound scan, or sonography, is a non-invasive, painless diagnostic procedure that has been used for several decades in medical practice. Also known as ultrasonography, this imaging technique uses high-frequency sound waves to create images of the developing fetus and its surrounding structures. These images provide information about the growth and development of the fetus inside the mother’s womb.

Why is Ultrasound Needed During Pregnancy?

Ultrasound is an important tool for monitoring the health and development of your baby during pregnancy. Additionally, it provides parents with the first glimpse of their growing little human!

The USG (ultrasound sonography test) is necessary to diagnose the health of the growing baby and to detect and diagnose potential problems such as ectopic pregnancy, multiple gestation, fetal anomalies, and placental abnormalities. The procedure also monitors the growth and development of the fetus and the amount of amniotic fluid surrounding the baby.

What are the Types of Ultrasound Performed During Pregnancy?

There are several types of ultrasound scans performed during pregnancy:

1.       Transabdominal Ultrasound: This is the most common technique, where a transducer is placed on the mother’s abdomen to produce images of the growing fetus.

2.       Transvaginal Ultrasound: This technique involves inserting a narrow transducer into the vagina to create images of the fetus, typically carried out very early during pregnancy for a clearer view.

3.       Doppler Ultrasound: This scan uses sound waves to assess blood flow in the placenta, umbilical cord, and fetus, frequently used throughout the second and third trimesters to check on fetal health and well-being.

4.       3D/4D Ultrasound: These specialized forms of ultrasound produce three-dimensional or four-dimensional images, giving parents a clearer view of the baby’s features.

5.       Fetal Echocardiography: This specialized procedure focuses on the developing fetus’s heart and is typically carried out when there are concerns about the baby’s heart health.

How is Ultrasound Beneficial During Pregnancy?

Beyond confirming the fact that you are pregnant, ultrasound scans offer several advantages, such as:

1.       Early Detection of Potential Problems: USG helps detect potential problems early, allowing for timely intervention and treatment.

2.       Detection of Ectopic Pregnancy: USG can detect ectopic pregnancies, a potentially life-threatening condition where the fertilized egg implants outside the uterus.

3.       Diagnosis of Abnormalities: Scans can detect fetal abnormalities, such as neural tube defects and heart defects. Nuchal translucency (NT) scans during pregnancy help detect chromosomal abnormalities.

4.       Assessment of Amniotic Fluid Levels: Sonography helps detect the levels of amniotic fluid around the fetus, which is crucial for monitoring fetal well-being.

5.       Detecting the Number of Babies: The procedure helps detect the number of babies (twins, triplets, etc.), requiring special monitoring during pregnancy.

6.       Confirmation of Due Date: Confirming your due date becomes easier with a USG procedure.

7.       Position of the Baby: The technique helps detect the baby’s position, aiding in delivery planning.

When is the First Ultrasound Suggested?

The Obstetricians and Gynecologists recommends that the first ultrasound scan be performed between 6-8 weeks of pregnancy. This is typically done to confirm the pregnancy and to check for the presence of a fetal heartbeat.

At Butler Medical Imaging, we understand the significance of these moments and are committed to providing exceptional ultrasound imaging services. With state-of-the-art technology and skilled technicians, we ensure high-quality images that capture the essence of your baby’s development, creating lasting memories for you and your family.

Why Choose Butler Medical Imaging?

We are a local, privately-owned medical clinic that offers the highest quality medical and imaging services in the Butler Medical Imaging region. We are leaders in bulk billing, providing the real benefit of no 'Out of Pocket' expenses for you.

Your Next Step

Please ensure that you have your Medicare card and referral with you and visit our location for your consultation. If you have any questions, please feel free to contact us at +61 8 9544 3999. We are here to help.